backend/requirements.txt
# DeepFake Guard – Python 3.11+ requirements
# CPU-only PyTorch for lightweight Docker/Render deployments
# To use GPU: replace torch/torchvision lines with CUDA variants from https://pytorch.org

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20      # required for UploadFile

# ── Config ────────────────────────────────────────────────────────────────────
pydantic-settings==2.9.1

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python-headless==4.11.0.86
Pillow==11.2.1

# ── Deep Learning (CPU-only) ─────────────────────────────────────────────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchvision==0.22.1+cpu

# ── Training utilities ────────────────────────────────────────────────────────
scikit-learn==1.6.1            # roc_auc_score — pre-built wheels for Python 3.13
numpy==2.2.3
tqdm==4.67.1                   # progress bars during training

# ── Dataset download (used by utils/download_dataset.py) ─────────────────────
datasets==3.6.0                # Hugging Face datasets library
huggingface_hub==0.32.4        # HF Hub API (authentication, file download)